Split disc track list into own file, unwrapped duration
